feat : TerminalBar View 분리 #11
